requirements.txt
# --------------------------------------------------------------------------------------------- # 
# | Name: Md. Shahanur Islam Shagor                                                           | # 
# | Autonomous Systems & UAV Researcher | Cybersecurity    | Specialist | Software Engineer   | #
# | Voronezh State University of Forestry and Technologies                                    | # 
# | Build for Blind people within 15$                                                         | # 
# --------------------------------------------------------------------------------------------- # 

# Computer Vision and Machine Learning
opencv-python
opencv-contrib-python
ultralytics
torch
torchvision
numpy

# Text-to-Speech
pyttsx3

# Network and HTTP
urllib3
requests

# Utilities
Pillow
pyyaml
psutil

# Security / Realtime Control
pycryptodome
websockets
paho-mqtt

# Accelerated Inference (Intel)
openvino
prometheus-client
